Made some adjustments in order to be compatible with the new version of XCSP3 API.
git-svn-id: svn://svn.forge.objectweb.org/svnroot/sat4j/maven/trunk@2548 ba638df5-4473-46d1-82f8-c3ae2a17a6e1

+++ org.sat4j.csp/src/main/java/org/sat4j/csp/constraints3/StringCondition.java
@@ -47,11 +47,11 @@ public class StringCondition {
 	}
 	
 	public StringCondition(ConditionVal condition) {
-		this.asString = condition.operator.name().toLowerCase()+"("+LEFT_OPERAND+","+Integer.toString(((ConditionVal)condition).k)+")";
+		this.asString = condition.operator.name().toLowerCase()+"("+LEFT_OPERAND+","+Long.toString(((ConditionVal)condition).k)+")";
 	}
 	
 	public StringCondition(ConditionIntvl condition) {
-		this.asString = "and(ge("+LEFT_OPERAND+","+Integer.toString(condition.min)+"),le("+LEFT_OPERAND+","+Integer.toString(condition.max)+"))";
+		this.asString = "and(ge("+LEFT_OPERAND+","+Long.toString(condition.min)+"),le("+LEFT_OPERAND+","+Long.toString(condition.max)+"))";
 	}
 	
 	public StringCondition(String str) {
